requirements.txt
razorpay
python-binance
requests
python-telegram-bot
pyyaml
fpdf
